Moved FlowIntegrationTestCase to the same namespace as FlowTestCase
