+++ src/adapter/etl-adapter-xml/tests/Flow/ETL/Adapter/XML/Tests/Integration/Loader/XMLLoaderTest.php
@@ -7,7 +7,7 @@ namespace Flow\ETL\Adapter\XML\Tests\Integration\Loader;
 use function Flow\ETL\Adapter\XML\{from_xml, to_xml};
 use function Flow\ETL\DSL\{df, from_array, overwrite, ref};
 use Flow\ETL\Tests\Double\FakeExtractor;
-use Flow\ETL\Tests\Integration\FlowIntegrationTestCase;
+use Flow\ETL\Tests\FlowIntegrationTestCase;
 
 final class XMLLoaderTest extends FlowIntegrationTestCase
 {
