+++ src/adapter/etl-adapter-xml/tests/Flow/ETL/Adapter/XML/Tests/Integration/XMLParserExtractorTest.php
@@ -12,7 +12,7 @@ use Flow\ETL\{Adapter\XML\XMLParserExtractor,
     Flow,
     FlowContext,
     PHP\Type\Caster,
-    Tests\Integration\FlowIntegrationTestCase};
+    Tests\FlowIntegrationTestCase};
 use Flow\Filesystem\Path;
 
 final class XMLParserExtractorTest extends FlowIntegrationTestCase
