+++ src/adapter/etl-adapter-xml/tests/Flow/ETL/Adapter/XML/Tests/Integration/XMLReaderExtractorTest.php
@@ -7,7 +7,7 @@ namespace Flow\ETL\Adapter\XML\Tests\Integration;
 use function Flow\ETL\DSL\type_string;
 use Flow\ETL\Adapter\XML\XMLReaderExtractor;
 use Flow\ETL\Extractor\Signal;
-use Flow\ETL\{Config, Flow, FlowContext, PHP\Type\Caster, Tests\Integration\FlowIntegrationTestCase};
+use Flow\ETL\{Config, Flow, FlowContext, PHP\Type\Caster, Tests\FlowIntegrationTestCase};
 use Flow\Filesystem\Path;
 
 final class XMLReaderExtractorTest extends FlowIntegrationTestCase
