+++ src/bridge/filesystem/async-aws/tests/Flow/Filesystem/Bridge/AsyncAWS/Tests/Integration/AsyncAWSS3TestCase.php
@@ -6,7 +6,7 @@ namespace Flow\Filesystem\Bridge\AsyncAWS\Tests\Integration;
 
 use function Flow\Filesystem\Bridge\AsyncAWS\DSL\aws_s3_client;
 use AsyncAws\S3\S3Client;
-use Flow\ETL\Tests\Integration\FlowIntegrationTestCase;
+use Flow\ETL\Tests\FlowIntegrationTestCase;
 use Flow\Filesystem\Path;
 
 abstract class AsyncAWSS3TestCase extends FlowIntegrationTestCase
