+++ src/core/etl/tests/Flow/ETL/Tests/Integration/DataFrame/CacheTest.php
@@ -8,7 +8,7 @@ use function Flow\ETL\DSL\{config_builder, df, from_array, from_cache};
 use Flow\ETL\Cache\CacheIndex;
 use Flow\ETL\Cache\Implementation\InMemoryCache;
 use Flow\ETL\Tests\Double\FakeExtractor;
-use Flow\ETL\Tests\Integration\FlowIntegrationTestCase;
+use Flow\ETL\Tests\FlowIntegrationTestCase;
 use Flow\ETL\{Extractor, FlowContext, Rows};
 
 final class CacheTest extends FlowIntegrationTestCase
