+++ src/core/etl/tests/Flow/ETL/Tests/Integration/DataFrame/FilterTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Flow\ETL\Tests\Integration\DataFrame;
 
 use function Flow\ETL\DSL\{df, from_array, ref};
-use Flow\ETL\Tests\Integration\FlowIntegrationTestCase;
+use Flow\ETL\Tests\FlowIntegrationTestCase;
 
 final class FilterTest extends FlowIntegrationTestCase
 {
