+++ src/core/etl/tests/Flow/ETL/Tests/Integration/DataFrame/RenameTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Flow\ETL\Tests\Integration\DataFrame;
 
 use function Flow\ETL\DSL\{bool_entry, df, from_rows, int_entry, json_entry, ref, str_entry};
-use Flow\ETL\Tests\Integration\FlowIntegrationTestCase;
+use Flow\ETL\Tests\FlowIntegrationTestCase;
 use Flow\ETL\Transformer\StyleConverter\StringStyles;
 use Flow\ETL\{Row, Rows};
