+++ src/core/etl/tests/Flow/ETL/Tests/Integration/Monitoring/Memory/ConfigurationTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Flow\ETL\Tests\Integration\Monitoring\Memory;
 
 use Flow\ETL\Monitoring\Memory\{Configuration, Unit};
-use Flow\ETL\Tests\Integration\FlowIntegrationTestCase;
+use Flow\ETL\Tests\FlowIntegrationTestCase;
 
 final class ConfigurationTest extends FlowIntegrationTestCase
 {
