+++ src/core/etl/tests/Flow/ETL/Tests/Integration/Sort/ExternalSort/ExternalSortTest.php
@@ -8,7 +8,7 @@ use function Flow\ETL\DSL\{flow_context, from_array, ref, refs};
 use Flow\ETL\Pipeline\{SynchronousPipeline};
 use Flow\ETL\Sort\ExternalSort;
 use Flow\ETL\Sort\ExternalSort\BucketsCache\FilesystemBucketsCache;
-use Flow\ETL\Tests\Integration\FlowIntegrationTestCase;
+use Flow\ETL\Tests\FlowIntegrationTestCase;
 use Flow\Filesystem\Path;
 
 final class ExternalSortTest extends FlowIntegrationTestCase
